+++ data-raw/ebd-samples.r
@@ -3,21 +3,24 @@ library(tidyverse)
 library(stringi)
 library(stringr)
 
-ebird_dir <- "~/data/ebird/ebd_relFeb-2018/"
+ebird_dir <- "~/data/ebird/ebd_relMay-2018/"
 # ebd
-f_in <- file.path(ebird_dir, "ebd_relFeb-2018.txt")
-f_subset <- file.path(ebird_dir, "ebd_relFeb-2018_subset.txt")
-f_sg <- file.path(ebird_dir, "ebd_relFeb-2018_SG.txt")
-f_ru <- file.path(ebird_dir, "ebd_relFeb-2018_rollup.txt")
+f_in <- file.path(ebird_dir, "ebd_relMay-2018.txt")
+f_subset <- file.path(ebird_dir, "ebd_relMay-2018_subset.txt")
+f_sg <- file.path(ebird_dir, "ebd_relMay-2018_SG.txt")
+f_ru <- file.path(ebird_dir, "ebd_relMay-2018_rollup.txt")
 # sampling
-s_in <- file.path(ebird_dir, "ebd_sampling_relFeb-2018.txt")
-s_subset <- file.path(ebird_dir, "ebd_sampling_relFeb-2018_subset.txt")
-s_sg <- file.path(ebird_dir, "ebd_sampling_relFeb-2018_SG.txt")
+s_in <- file.path(ebird_dir, "ebd_sampling_relMay-2018.txt")
+s_subset <- file.path(ebird_dir, "ebd_sampling_relMay-2018_subset.txt")
+s_sg <- file.path(ebird_dir, "ebd_sampling_relMay-2018_SG.txt")
 
 # filter
 filters <- auk_ebd(f_in) %>%
-  auk_species(species = c("Gray Jay", "Blue Jay", "Steller's Jay", "Green Jay")) %>%
-  auk_country(country = c("US", "Canada", "Mexico", "Belize", "Guatemala", "Honduras", "Panama", "Costa Rica", "El Salvador")) %>%
+  auk_species(species = c("Gray Jay", "Blue Jay", 
+                          "Steller's Jay", "Green Jay")) %>%
+  auk_country(country = c("US", "Canada", "Mexico", "Belize", 
+                          "Guatemala", "Honduras", "Panama", 
+                          "Costa Rica", "El Salvador")) %>%
   auk_date(date = c("2010-01-01", "2012-12-31")) %>%
   auk_time(start_time = c("06:00", "12:00")) %>%
   auk_duration(duration = c(0, 120))
@@ -27,7 +30,7 @@ if (!file.exists(f_subset)) {
 
 x <- read_tsv(f_subset, quote = "", 
               col_types = cols(.default = col_character())) %>% 
-  select(-X49)
+  select(-X47)
 # evenly sample species
 set.seed(1)
 n_min <- min(table(x$`SCIENTIFIC NAME`))
@@ -38,10 +41,7 @@ x <- x %>%
 # sample to 500 records, make sure to get some from central america
 y1 <- sample_n(x %>% filter(!`COUNTRY CODE` %in% c("CA", "US")), 100)
 y2 <- sample_n(x %>% filter(`COUNTRY CODE` %in% c("CA", "US")), 400)
-y <- bind_rows(y1, y2) %>% 
-  # fill with fake names for privacy
-  mutate(`FIRST NAME` = stri_trans_general(`FIRST NAME`, "latin-ascii"),
-         `LAST NAME` = "eBirder")
+y <- bind_rows(y1, y2)
 f <- "inst/extdata/ebd-sample.txt"
 write_tsv(y, f, na = "")
 # remove any non-ascii characters
@@ -68,7 +68,8 @@ stopifnot(all(read_ebd(f)$scientific_name %in% ebird_taxonomy$scientific_name))
 
 # filter for zero-fill example
 filters <- auk_ebd(f_in, s_in) %>%
-  auk_species(species = c("Collared Kingfisher", "White-throated Kingfisher", "Blue-eared Kingfisher")) %>%
+  auk_species(species = c("Collared Kingfisher", "White-throated Kingfisher", 
+                          "Blue-eared Kingfisher")) %>%
   auk_country(country = "Singapore") %>%
   auk_date(date = c("2012-01-01", "2012-12-31"))
 if (!file.exists(f_sg)) {
@@ -79,9 +80,7 @@ if (!file.exists(f_sg)) {
 # observations
 x_ebd <- read_tsv(f_sg, quote = "", 
                   col_types = cols(.default = col_character())) %>% 
-  select(-X49) %>% 
-  mutate(`FIRST NAME` = stri_trans_general(`FIRST NAME`, "latin-ascii"),
-         `LAST NAME` = "eBirder")
+  select(-X47)
 f <- "inst/extdata/zerofill-ex_ebd.txt"
 write_tsv(x_ebd, f, na = "")
 # remove any non-ascii characters
@@ -94,9 +93,7 @@ stopifnot(length(tools::showNonASCII(readLines(f))) == 0)
 stopifnot(all(read_ebd(f)$scientific_name %in% ebird_taxonomy$scientific_name))
 # checklists
 x_samp <- read_tsv(s_sg, quote = "", col_types = cols(.default = col_character())) %>% 
-  select(-X33) %>% 
-  mutate(`FIRST NAME` = stri_trans_general(`FIRST NAME`, "latin-ascii"),
-         `LAST NAME` = "eBirder")
+  select(-X31)
 f <- "inst/extdata/zerofill-ex_sampling.txt"
 write_tsv(x_samp, f, na = "")
 # remove any non-ascii characters
@@ -109,21 +106,28 @@ stopifnot(length(tools::showNonASCII(readLines(f))) == 0)
 
 # rollup
 if (!file.exists(f_ru)) {
-  paste("head -1", f_in, ">", f_ru) %>% 
+  paste("head -1000000", f_in, ">", f_ru) %>% 
     system()
   paste("tail -1000000", f_in, ">>", f_ru) %>% 
     system()
 }
 x_ebd <- read_tsv(f_ru, quote = "", col_types = cols(.default = col_character())) %>% 
-  select(-X49) %>% 
-  mutate(`FIRST NAME` = stri_trans_general(`FIRST NAME`, "latin-ascii"),
-         `LAST NAME` = "eBirder")
+  select(-X47)
+yrwa_all <- x_ebd %>% 
+  filter(`COMMON NAME` == "Yellow-rumped Warbler") %>% 
+  group_by(`SAMPLING EVENT IDENTIFIER`) %>% 
+  count() %>% 
+  filter(n == 3) %>% 
+  pull(`SAMPLING EVENT IDENTIFIER`) %>% 
+  sample(1)
+stopifnot(length(yrwa_all) == 1)
 # yellow-rumped
 set.seed(1)
-ru_ex <- filter(x_ebd, `SAMPLING EVENT IDENTIFIER` == "S41507433", 
+ru_ex <- filter(x_ebd, `SAMPLING EVENT IDENTIFIER` == yrwa_all, 
                 `COMMON NAME` == "Yellow-rumped Warbler")
 ru_ex <- x_ebd %>% 
-  filter(CATEGORY %in% c("spuh", "slash", "hybrid", "domestic", "form", "intergrade")) %>% 
+  filter(CATEGORY %in% c("spuh", "slash", "hybrid", "domestic", "form", 
+                         "intergrade")) %>% 
   group_by(CATEGORY) %>% 
   sample_n(3) %>% 
   ungroup() %>% 
@@ -140,6 +144,7 @@ stopifnot(length(tools::showNonASCII(readLines(f))) == 0)
 stopifnot(all(read_ebd(f)$scientific_name %in% ebird_taxonomy$scientific_name))
 
 # after script: edit the messy file to introduce errors, especially tabs in comment fields
+# keep top 5 rows intact
 f <- "inst/extdata/ebd-sample_messy.txt"
 stopifnot(length(tools::showNonASCII(readLines(f))) == 0)
 stopifnot(all(read_ebd(f, reader = "base")$scientific_name %in% ebird_taxonomy$scientific_name))